+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/TrainModelWithoutLabelsAsync.java
@@ -43,7 +43,7 @@ public class TrainModelWithoutLabelsAsync {
         PollerFlux<FormRecognizerOperationResult, CustomFormModel> trainingPoller
             = client.beginTraining(trainingFilesUrl,
             false,
-            new TrainingOptions().setModelDisplayName("composed model name"));
+            new TrainingOptions().setModelName("composed model name"));
 
         Mono<CustomFormModel> customFormModelResult = trainingPoller
             .last()
@@ -61,7 +61,7 @@ public class TrainModelWithoutLabelsAsync {
             // Model Info
             System.out.printf("Model Id: %s%n", customFormModel.getModelId());
             System.out.printf("Model Status: %s%n", customFormModel.getModelStatus());
-            System.out.printf("Model display name: %s%n", customFormModel.getModelDisplayName());
+            System.out.printf("Model display name: %s%n", customFormModel.getModelName());
             System.out.printf("Training started on: %s%n", customFormModel.getTrainingStartedOn());
             System.out.printf("Training completed on: %s%n%n", customFormModel.getTrainingCompletedOn());